requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosmtplib==4.0.2
alembic==1.17.0
amqp==2.6.1
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
atomicwrites==1.4.1
attrs==25.4.0
bcrypt==4.3.0
billiard==3.6.4.0
bitarray==3.7.2
cachetools==6.2.1
celery==4.4.7
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
ckzg==2.1.5
click==7.1.2
colorama==0.4.6
cryptography==46.0.3
cssselect==1.3.0
cssutils==2.11.1
cytoolz==1.1.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator>=2.0.0
emails==0.5.15
eth-account==0.10.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.115.14
frozenlist==1.8.0
greenlet==3.2.4
gunicorn==20.1.0
h11==0.16.0
hexbytes==0.3.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
inflect==5.6.2
iniconfig==2.3.0
Jinja2==2.11.3
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==4.6.11
lru-dict==1.2.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
mcp==1.18.0
more-itertools==10.8.0
multidict==6.7.0
openai==2.6.0
orjson==3.10.18
packaging==25.0
parsimonious==0.10.0
passlib==1.7.4
pluggy==1.6.0
premailer==3.10.0
propcache==0.4.1
protobuf==6.33.0
psycopg2-binary==2.9.11
py==1.11.0
pyasn1==0.6.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pyunormalize==17.0.0
raven==6.10.0
redis==6.2.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rlp==4.1.0
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sse-starlette==3.0.2
starlette==0.46.2
tenacity==6.3.1
toolz==1.1.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.34.3
vine==1.3.0
wcwidth==0.2.14
web3==6.15.1
websockets==15.0.1
yarl==1.22.0